# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: M-ALDBANI platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /projects:
    get:
      operationId: listProjects
      tags: [public]
      summary: List portfolio projects
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [public]
      summary: Get a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  # PUBLIC - SERVICES
...
  /articles:
    get:
      operationId: listArticles
      tags: [public]
      summary: List articles
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
  /articles/{slug}:
    get:
      operationId: getArticle
      tags: [public]
      summary: Get an article by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"

  # PUBLIC - CONTACT & BOOKING
  /contact:
    post:
      operationId: submitContact
      tags: [public]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/leads:
    get:
      operationId: listLeads
      tags: [admin]
      summary: List all leads
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [admin]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /admin/leads/{id}:
    patch:
      operationId: updateLead
      tags: [admin]
      summary: Update a lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [admin]
      summary: Delete a lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/consultations:
    get:
      operationId: listAllConsultations
      tags: [admin]
      summary: List all consultations
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultation"
  /admin/consultations/{id}:
    patch:
      operationId: updateConsultation
      tags: [admin]
      summary: Update consultation status/notes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /admin/projects:
    get:
      operationId: listAdminProjects
      tags: [admin]
      summary: List all projects (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [admin]
      summary: Create a portfolio project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /admin/projects/{id}:
    patch:
      operationId: updateProject
      tags: [admin]
      summary: Update a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [admin]
      summary: Delete a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/articles:
    post:
      operationId: createArticle
      tags: [admin]
      summary: Create an article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /admin/articles/{id}:
    patch:
      operationId: updateArticle
      tags: [admin]
      summary: Update an article
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    delete:
      operationId: deleteArticle
      tags: [admin]
      summary: Delete an article
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/services:
    post:
      operationId: createService
      tags: [admin]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /admin/services/{id}:
    patch:
      operationId: updateService
      tags: [admin]
      summary: Update a service
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [admin]
      summary: Delete a service
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        avatar:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, email, role, createdAt]
    RegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
      required: [name, email, password]
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    Project:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        category:
          type: string
        technologies:
          type: array
          items:
            type: string
        image:
          type: string
        images:
          type: array
          items:
            type: string
        results:
          type: ["string", "null"]
        featured:
          type: boolean
        order:
          type: integer
        createdAt:
          type: string
      required: [id, title, titleAr, description, descriptionAr, category, technologies, image, featured, order, createdAt]
    ProjectInput:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        category:
          type: string
        technologies:
          type: array
          items:
            type: string
        image:
          type: string
        results:
          type: string
        featured:
          type: boolean
        order:
          type: integer
      required: [title, titleAr, description, descriptionAr, category, technologies, image]
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        category:
          type: string
        technologies:
          type: array
          items:
            type: string
        image:
          type: string
        results:
          type: string
        featured:
          type: boolean
        order:
          type: integer
    Service:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        icon:
          type: string
        price:
          type: ["string", "null"]
        featured:
          type: boolean
        order:
          type: integer
      required: [id, title, titleAr, description, descriptionAr, icon, featured, order]
    ServiceInput:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        icon:
          type: string
        price:
          type: string
        featured:
          type: boolean
        order:
          type: integer
      required: [title, titleAr, description, descriptionAr, icon]
    ServiceUpdate:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        icon:
          type: string
        price:
          type: string
        featured:
          type: boolean
        order:
          type: integer
    Article:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        titleAr:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        excerptAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        category:
          type: string
        coverImage:
          type: string
        readTime:
          type: integer
        published:
          type: boolean
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, title, titleAr, slug, excerpt, excerptAr, content, contentAr, category, coverImage, readTime, published, createdAt]
    ArticleInput:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        excerptAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        category:
          type: string
        coverImage:
          type: string
        readTime:
          type: integer
        published:
          type: boolean
      required: [title, titleAr, slug, excerpt, excerptAr, content, contentAr, category, coverImage]
    ArticleUpdate:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        excerpt:
          type: string
        excerptAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        category:
          type: string
        coverImage:
          type: string
        readTime:
          type: integer
        published:
          type: boolean
    Consultation:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: ["string", "null"]
        clientName:
          type: string
        clientEmail:
          type: string
        clientPhone:
          type: string
        type:
          type: string
        date:
          type: string
        time:
          type: string
        duration:
          type: integer
        status:
          type: string
        notes:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        meetingLink:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        paid:
          type: boolean
        createdAt:
          type: string
      required: [id, clientName, clientEmail, clientPhone, type, date, time, duration, status, paid, createdAt]
...
    ConsultationUpdate:
      type: object
      properties:
        status:
          type: string
        adminNotes:
          type: string
        meetingLink:
          type: string
        price:
          type: number
        paid:
          type: boolean
    ContactInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        message:
          type: string
        subject:
          type: string
      required: [name, email, message]
    Lead:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        source:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        createdAt:
          type: string
      required: [id, name, email, source, status, createdAt]
    LeadInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        source:
          type: string
        status:
          type: string
        notes:
          type: string
        value:
          type: number
      required: [name, email, source]
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        source:
          type: string
        status:
          type: string
        notes:
          type: string
        value:
          type: number
...